requirements.txt
streamlit>=1.30.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
edgartools>=2.0.0
pydantic>=2.0.0
